requirements.txt
moviepy==1.0.3
imageio-ffmpeg>=0.4.8
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
openai>=1.40.0
supabase>=2.4.0
edge-tts==0.1.1
